backend/requirements.txt
# Repo Auditor Backend - Runtime Dependencies

# Web framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database
sqlalchemy>=2.0.0
alembic>=1.13.0
asyncpg>=0.29.0
aiosqlite>=0.19.0

# Git operations
gitpython>=3.1.40

# Code analysis
semgrep>=1.50.0
radon>=6.0.0

# HTTP client
httpx>=0.26.0

# Security
pyjwt>=2.8.0

# Config & Logging
pyyaml>=6.0.1
structlog>=24.1.0

# Report generation
jinja2>=3.1.0
markdown>=3.5.0

# Export formats (optional)
openpyxl>=3.1.0
xhtml2pdf>=0.2.11

# Google Drive integration
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0

# LLM Providers (install as needed)
anthropic>=0.18.0
openai>=1.12.0
google-generativeai>=0.3.0
